refactor: PromptModal 컴포넌트에서 Modal 컴포넌트 사용
